requirements.txt
ollama
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
discord.py
flask
flask-cors
requests
